Fix warning: Expression 'nVisiblePart >= 0' is always true. Unsigned type value is always >= 0.

+++ src/ConEmu/ConEmu.cpp
@@ -8237,7 +8237,7 @@ UINT CConEmuMain::IsQuakeVisible()
 				_ASSERTE(nVisiblePart <= 100);
 				nVisiblePart = 100;
 			}
-			_ASSERTE(nVisiblePart >= 0);
+			_ASSERTE((int)nVisiblePart >= 0);
 		}
 		else
 		{
